+++ packages/components/src/systems/RecentOrders/RecentOrdersDuplicateDialog.tsx
@@ -1,19 +1,31 @@
+import { createEffect, Ref } from 'solid-js';
 import { useRecentOrders } from '.';
 import Button from '../../particles/Button';
 import Dialog from '../../particles/Dialog';
 import Icon from '../../particles/Icon';
 import Text from '../../particles/Text';
+import { dispatchDatadogAction } from '../../utils/dispatchDatadogAction';
 import formatRxString from '../../utils/formatRxString';
 
 export default function RecentOrdersDuplicateDialog() {
+  let ref: Ref<any> | undefined;
   const [state, actions] = useRecentOrders();
+
+  createEffect(() => {
+    if (state.isDuplicateDialogOpen) {
+      dispatchDatadogAction('prescribe-duplicate-dialog-open', {}, ref);
+    }
+  });
 
   return (
     <Dialog
       open={state.isDuplicateDialogOpen}
-      onClose={() => actions.setIsDuplicateDialogOpen(false)}
+      onClose={() => {
+        dispatchDatadogAction('prescribe-duplicate-dialog-exit', {}, ref);
+        actions.setIsDuplicateDialogOpen(false);
+      }}
     >
-      <div class="flex flex-col gap-6">
+      <div class="flex flex-col gap-6" ref={ref}>
         <div>
           <div class="table bg-red-50 text-red-600 p-2 rounded-full mb-4">
             <Icon name="exclamationCircle" />
@@ -51,13 +63,21 @@ export default function RecentOrdersDuplicateDialog() {
             size="xl"
             variant="secondary"
             onClick={() => {
+              dispatchDatadogAction('prescribe-duplicate-dialog-add-anyway', {}, ref);
               state?.duplicateDialogContinueCb?.();
               actions.setIsDuplicateDialogOpen(false);
             }}
           >
             Add Prescriptions Anyway
           </Button>
-          <Button variant="naked" size="xl" onClick={() => actions.setIsDuplicateDialogOpen(false)}>
+          <Button
+            variant="naked"
+            size="xl"
+            onClick={() => {
+              dispatchDatadogAction('prescribe-duplicate-dialog-exit', {}, ref);
+              actions.setIsDuplicateDialogOpen(false);
+            }}
+          >
             Go Back
           </Button>
         </div>
