+++ packages/components/src/utils/dispatchDatadogAction.ts
@@ -1,0 +1,17 @@
+export const dispatchDatadogAction = (
+  action: string,
+  data: {
+    [key: string]: unknown;
+  },
+  ref?: HTMLElement
+) => {
+  const event = new CustomEvent('photon-datadog-action', {
+    composed: true,
+    bubbles: true,
+    detail: {
+      action,
+      data
+    }
+  });
+  ref?.dispatchEvent(event);
+};
